requirements.txt
pyyaml~=6.0.1
torch~=2.0.0+cu117